Resize modal and add vertical border lines
Made the overall DialogContent more like the shape of the Figma, and then added some vertical border lines, unfortunately these are 2 pixels thick, and disappear when they are made 1px, so looks a bit weird. Apart from that, removed the x-axis scroll bar as it is unlikely to be used.